+++ tensorcircuit/templates/__init__.py
@@ -6,5 +6,6 @@ from . import graphs
 from . import measurements
 from . import conversions
 from . import lattice
+from . import hamiltonians
 
 costfunctions = measurements